+++ src/types.rs
@@ -4,9 +4,10 @@
 
 //! Generic types.
 
-use std::collections::hash_map::DefaultHasher;
-use std::default::Default;
-use std::hash::{Hash, Hasher};
+use std::{
+    collections::hash_map::DefaultHasher,
+    hash::{Hash, Hasher},
+};
 
 #[derive(Debug, Clone, Copy)]
 pub(crate) enum Pol {